This commit was generated by cvs2svn to compensate for changes in r8297, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@8298 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/vicii/vicii-mem.c
@@ -1047,7 +1047,7 @@ void REGPARM2 vic_store(ADDRESS addr, BYTE value)
     case 0xb:                   /* $D00B: Sprite #5 Y position */
     case 0xd:                   /* $D00D: Sprite #6 Y position */
     case 0xf:                   /* $D00F: Sprite #7 Y position */
-      store_sprite_y_position(addr, (value + vic_ii.offset) & 255);
+      store_sprite_y_position(addr, (BYTE)((value + vic_ii.offset) & 255));
       break;
 
     case 0x10:                  /* $D010: Sprite X position MSB */
